Move update_singleton() outside of testing utils
