+++ tests/unittests/storage/test_legacy.py
@@ -10,9 +10,10 @@ import pytest
 
 from orion.core.io.database import Database
 from orion.core.io.database.pickleddb import PickledDB
-from orion.core.utils import SingletonAlreadyInstantiatedError, SingletonNotInstantiatedError
+from orion.core.utils import SingletonAlreadyInstantiatedError, SingletonNotInstantiatedError, \
+    update_singletons
 from orion.core.utils.exceptions import MissingResultFile
-from orion.core.utils.tests import OrionState, update_singletons
+from orion.core.utils.tests import OrionState
 from orion.core.worker.trial import Trial
 from orion.storage.base import FailedUpdate
 from orion.storage.legacy import get_database, setup_database
